move example script from test to example folder
